Integrated API Searches
> Added axios dependency for run time fetching. This will be the method
for pulling meal results from the API.
> Added fetchMeal() method to index.js which searches the STC API based on
parameters saved in the local state.
> Changed filteredCategories and filteredOrigins state variables and
update methods to use strings instead of id numbers. This is due to how
the API is currently setup to take in strings for parameters and not the
ids.
> Updated getMealButton to use FetchMeal() method when clicked.
> Removed setTimeout from togglefilteredCategories and
toggleFilteredOrigins methods as it was causing issues when clicking
buttons too quickly.